updated test files for all event related screens, plus unit testing file for DateFormatter class
