+++ app/src/androidTest/java/com/example/butter/CreateEventTest.java
@@ -41,6 +41,17 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+/**
+ * NOTE: IN ORDER FOR THESE TESTS TO WORK, YOU HAVE TO FIRST LAUNCH THE APP NORMALLY (NOT THE TEST FILE)
+ *       ONCE IT'S BOOTED UP FULLY, CLICK THE DROPDOWN BUTTON TO THE LEFT OF THE RUN APP BUTTON
+ *       CLICK RUN 'EVENT SCREEN TEST'
+ *
+ *       IF YOU TRY TO RUN THE TEST FILE DIRECTLY WITHOUT THESE STEPS, IT WILL ALWAYS FAIL
+ *
+ *       IF THIS TRICK DOES NOT WORK ON YOUR MACHINE, PLEASE ASK ME TO SHOW YOU ON MY MACHINE
+ *
+ * @author Nate Pane (natepane)
+ */
 @RunWith(AndroidJUnit4.class)
 @LargeTest
 public class CreateEventTest {
