+++ app/src/androidTest/java/com/example/butter/EventScreenTest.java
@@ -1,36 +1,34 @@
 package com.example.butter;
 
-import static androidx.test.espresso.Espresso.onData;
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.clearText;
 import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.action.ViewActions.typeText;
-import static
-        androidx.test.espresso.assertion.ViewAssertions.doesNotExist;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.isAssignableFrom;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withSpinnerText;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.CoreMatchers.allOf;
 import static org.hamcrest.CoreMatchers.anything;
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
-import static org.hamcrest.CoreMatchers.startsWith;
 import static org.hamcrest.Matchers.hasToString;
 
-import androidx.test.core.app.ActivityScenario;
-import androidx.test.espresso.action.ViewActions;
-import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import android.content.Intent;
+import android.widget.DatePicker;
+
+import androidx.test.espresso.contrib.PickerActions;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
+import androidx.test.rule.ActivityTestRule;
 
-import org.hamcrest.Matcher;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
 
 /**
  * NOTE: IN ORDER FOR THESE TESTS TO WORK, YOU HAVE TO FIRST LAUNCH THE APP NORMALLY (NOT THE TEST FILE)
@@ -41,7 +39,7 @@ import org.junit.runner.RunWith;
  *
  *       IF THIS TRICK DOES NOT WORK ON YOUR MACHINE, PLEASE ASK ME TO SHOW YOU ON MY MACHINE
  *
- *       ALSO, YOU MUST HAVE AT LEAST ONE EVENT CREATED FOR THESE TESTS TO WORK
+ *       ALSO, FILL IN YOUR DEVICE ID, AND AN EVENT ID OF AN EVENT YOU ORGANIZE IN THE setUp METHOD
  *
  * @author Nate Pane (natepane)
  */
@@ -49,72 +47,41 @@ import org.junit.runner.RunWith;
 @LargeTest
 public class EventScreenTest {
     @Rule
-    public ActivityScenarioRule<MainActivity> scenario = new ActivityScenarioRule<MainActivity>(MainActivity.class);
+    public ActivityTestRule<EventDetailsActivity> activityRule =
+            new ActivityTestRule<>(EventDetailsActivity.class, false, false); // Don't launch automatically
 
-    @Test
-    public void testAddEventButton() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withText("My Events")).check(matches(isDisplayed()));
-        onView(withId(R.id.addButton)).perform(click());
-        onView(withText("Create Event")).check(matches(isDisplayed()));
-    }
+    @Before
+    public void setUp() {
+        String deviceID = "a256a5d278042a1d"; // CHANGE THIS TO YOUR DEVICE ID
+        String eventID = "Bowling-a256a5d278042a1d"; // CHANGE THIS TO THE ID OF AN EVENT YOU ORGANIZE
 
-    @Test
-    public void testInvalidEventDetailsNoDescription() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.addButton)).perform(click());
-        onView(withId(R.id.name_event)).perform(ViewActions.typeText("Test Event"));
-        onView(withId(R.id.start_date)).perform(ViewActions.typeText("2024-12-01"));
-        onView(withId(R.id.end_date)).perform(ViewActions.typeText("2024-12-02"));
-        onView(withId(R.id.event_date)).perform(ViewActions.typeText("2024-12-03"));
+        Intent intent = new Intent();
+        intent.putExtra("deviceID", deviceID);
+        intent.putExtra("eventID", eventID);
 
-        onView(withId(R.id.create_event_button)).perform(click());
-        onView(withText("Create Event")).check(matches(isDisplayed()));
-    }
-
-    @Test
-    public void testInvalidEventNotConsecutiveDates() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.addButton)).perform(click());
-        onView(withId(R.id.name_event)).perform(ViewActions.typeText("Test Event"));
-        onView(withId(R.id.start_date)).perform(ViewActions.typeText("2024-12-01"));
-        onView(withId(R.id.end_date)).perform(ViewActions.typeText("2024-11-30"));
-        onView(withId(R.id.event_date)).perform(ViewActions.typeText("2024-11-29"));
-        onView(withId(R.id.event_description)).perform(ViewActions.typeText("Event description"));
-
-        onView(withId(R.id.create_event_button)).perform(click());
-        onView(withText("Create Event")).check(matches(isDisplayed()));
+        activityRule.launchActivity(intent);
     }
 
     @Test
     public void testEventDetails() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
         onView(withText("Delete Event")).check(matches(isDisplayed()));
     }
 
     @Test
-    public void testOrganizerOptionsDialog() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
+    public void testOrganizerOptions() {
         onView(withId(R.id.organizer_opt_button)).perform(click());
-
         onView(withText("Organizer Options")).check(matches(isDisplayed()));
     }
 
     @Test
-    public void testEditEventButton() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
+    public void testEditEventScreen() {
         onView(withId(R.id.organizer_opt_button)).perform(click());
         onView(withId(R.id.edit_event_text)).perform(click());
         onView(withId(R.id.edit_event_button)).check(matches(isDisplayed()));
     }
 
     @Test
-    public void testEditDetailsInvalidDescription() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
+    public void testIllegalDescription() {
         onView(withId(R.id.organizer_opt_button)).perform(click());
         onView(withId(R.id.edit_event_text)).perform(click());
         onView(withId(R.id.event_description)).perform(clearText());
@@ -123,57 +90,125 @@ public class EventScreenTest {
     }
 
     @Test
-    public void testEditDetailsInvalidDate() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
+    public void testIEditEventNonConsecutiveDates() {
         onView(withId(R.id.organizer_opt_button)).perform(click());
         onView(withId(R.id.edit_event_text)).perform(click());
-        onView(withId(R.id.event_start_date)).perform(clearText());
-        onView(withId(R.id.event_start_date)).perform(ViewActions.typeText("1900-01-01"));
+        onView(withId(R.id.event_start_date)).perform(click());
+        onView(isAssignableFrom(DatePicker.class)).perform(PickerActions.setDate(2026, 1, 1));
+        onView(withText("OK")).perform(click());
         onView(withId(R.id.edit_event_button)).perform(click());
         onView(withId(R.id.edit_event_button)).check(matches(isDisplayed()));
     }
 
     @Test
-    public void testEditDetailsValid() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
+    public void testEditEventIllegalStartDate() {
         onView(withId(R.id.organizer_opt_button)).perform(click());
         onView(withId(R.id.edit_event_text)).perform(click());
+        onView(withId(R.id.event_start_date)).perform(click());
+        onView(isAssignableFrom(DatePicker.class)).perform(PickerActions.setDate(2024, 11, 1));
+        onView(withText("OK")).perform(click());
+
+        onView(withId(R.id.end_date)).perform(click());
+        onView(isAssignableFrom(DatePicker.class)).perform(PickerActions.setDate(2024, 11, 2));
+        onView(withText("OK")).perform(click());
+
+        onView(withId(R.id.event_date)).perform(click());
+        onView(isAssignableFrom(DatePicker.class)).perform(PickerActions.setDate(2024, 11, 3));
+        onView(withText("OK")).perform(click());
+
+        onView(withId(R.id.edit_event_button)).perform(click());
+        onView(withId(R.id.edit_event_button)).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void testEditEventValidDetails() {
+        onView(withId(R.id.organizer_opt_button)).perform(click());
+        onView(withId(R.id.edit_event_text)).perform(click());
+
+        onView(withId(R.id.event_date)).perform(click());
+        onView(isAssignableFrom(DatePicker.class)).perform(PickerActions.setDate(2025, 1, 1));
+        onView(withText("OK")).perform(click());
+
         onView(withId(R.id.event_description)).perform(clearText());
         onView(withId(R.id.event_description)).perform(typeText("New event description"));
+
         onView(withId(R.id.edit_event_button)).perform(click());
-        onView(withText("Ok")).perform(click());
+        onView(withText("Delete Event")).check(matches(isDisplayed()));
+
         onView(withText("New event description")).check(matches(isDisplayed()));
     }
 
     @Test
-    public void testViewEntrantsButton() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
+    public void testViewEntrants() {
         onView(withId(R.id.organizer_opt_button)).perform(click());
         onView(withId(R.id.view_entrants_text)).perform(click());
         onView(withText("Entrants list")).check(matches(isDisplayed()));
     }
 
     @Test
-    public void testEntrantsListDissapearingButton() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
+    public void testWaitlistButtons() {
         onView(withId(R.id.organizer_opt_button)).perform(click());
         onView(withId(R.id.view_entrants_text)).perform(click());
         onView(withId(R.id.generate_entrants_button)).check(matches(isDisplayed()));
+        onView(withId(R.id.sample_size)).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void testDrawListButtons() {
+        onView(withId(R.id.organizer_opt_button)).perform(click());
+        onView(withId(R.id.view_entrants_text)).perform(click());
         onView(withId(R.id.entrants_spinner)).perform(click());
         onView(withText("Draw")).perform(click());
         onView(withId(R.id.generate_entrants_button)).check(matches(not(isDisplayed())));
+        onView(withId(R.id.delete_entrant_button)).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void testRegisteredListNoButtons() {
+        onView(withId(R.id.organizer_opt_button)).perform(click());
+        onView(withId(R.id.view_entrants_text)).perform(click());
+        onView(withId(R.id.entrants_spinner)).perform(click());
+        onView(withText("Registered")).perform(click());
+        onView(withId(R.id.generate_entrants_button)).check(matches(not(isDisplayed())));
+        onView(withId(R.id.delete_entrant_button)).check(matches(not(isDisplayed())));
+    }
+
+    @Test
+    public void testCancelledListButtons() {
+        onView(withId(R.id.organizer_opt_button)).perform(click());
+        onView(withId(R.id.view_entrants_text)).perform(click());
+        onView(withId(R.id.entrants_spinner)).perform(click());
+        onView(withText("Cancelled")).perform(click());
+        onView(withId(R.id.generate_entrants_button)).check(matches(not(isDisplayed())));
+        onView(withId(R.id.delete_entrant_button)).check(matches(not(isDisplayed())));
+        onView(withId(R.id.draw_replacements_button)).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void testNotifications() {
+        onView(withId(R.id.organizer_opt_button)).perform(click());
+        onView(withId(R.id.send_notifications_text)).perform(click());
+        onView(withText("Event Announcement")).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void testNotificationsNoMessage() {
+        onView(withId(R.id.organizer_opt_button)).perform(click());
+        onView(withId(R.id.send_notifications_text)).perform(click());
+        onView(withId(R.id.send_button)).perform(click());
+        onView(withText("Event Announcement")).check(matches(isDisplayed()));
     }
 
     @Test
     public void testShowDetailsCode() {
-        onView(withId(R.id.eventsIcon)).perform(click());
-        onView(withId(R.id.events_list)).perform(click());
         onView(withId(R.id.organizer_opt_button)).perform(click());
         onView(withId(R.id.show_details_code_text)).perform(click());
         onView(withText("Details QR Code")).check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void testDeleteEventConfirmationDialogue() {
+        onView(withText("Delete Event")).perform(click());
+        onView(withText("Delete")).check(matches(isDisplayed()));
     }
 }
